Sort the 'insert_resources' data to make static_loader! deterministic
Currently, the `TokenStream` produced by the `static_loader!` macro
depends on the default `HashMap` iteration order, as well as the
filesystem directory iteration order.

To allow `static_loader!` to be compatible with a reproducible build,
I've sorted the output of `insert_resources` before iterating over it.
This ensures that the final `TokenStream` depends only on the loaded
filenames and the file contents.